refactor: 메세지 생성하는 로직 분리
